Add utils module and move logging capabilities to it
